createLineDto extracted
