+++ domain/src/main/java/be/xplore/notifyme/services/ILineService.java
@@ -1,7 +1,6 @@
 package be.xplore.notifyme.services;
 
 import be.xplore.notifyme.domain.Line;
-import be.xplore.notifyme.dto.CreateLineDto;
 import java.security.Principal;
 import java.util.List;
 
@@ -11,7 +10,8 @@ public interface ILineService {
 
   Line getLine(long lineId);
 
-  Line createLine(CreateLineDto createLineDto, Principal principal);
+  Line createLine(String note, int requiredStaff, long facilityId, long eventId,
+                  Principal principal);
 
   List<Line> getAvailableLinesForUser(String userId);
 }
